chore(internal/librarian): move helper methods for clean to command.go
